created server/function/checkforwinnner
